Add default Cytology Program and Other fixes
