+++ frontend/src/components/admin/userManagement/UserAddModify.js
@@ -118,14 +118,16 @@ function UserAddModify() {
       setIsLoading(true);
     } else {
       setUserData(res);
-      setValidation({
-        validatepassword: true,
-        password: true,
-        password2: true,
-        loginName: true,
-        firstName: true,
-        secondName: true,
-      });
+      if (res.loginUserId) {
+        setValidation({
+          validatepassword: true,
+          password: true,
+          password2: true,
+          loginName: true,
+          firstName: true,
+          secondName: true,
+        });
+      }
       var KeyList = [];
       Object.keys(res.selectedTestSectionLabUnits).map((key) =>
         KeyList.push(key),
@@ -377,7 +379,7 @@ function UserAddModify() {
     const value = e.target.value.trim();
     const isValid = loginNameRegex.test(value);
 
-    if (value && !isValid) {
+    if (!value || (value && !isValid)) {
       if (!notificationVisible) {
         setNotificationVisible(true);
         addNotification({
@@ -484,7 +486,7 @@ function UserAddModify() {
     const value = e.target.value;
     const isValid = nameRegex.test(value);
 
-    if (value && !isValid) {
+    if (!value || (value && !isValid)) {
       if (!notificationVisible) {
         setNotificationVisible(true);
         addNotification({
@@ -517,7 +519,7 @@ function UserAddModify() {
     const value = e.target.value;
     const isValid = nameRegex.test(value);
 
-    if (value && !isValid) {
+    if (!value || (value && !isValid)) {
       if (!notificationVisible) {
         setNotificationVisible(true);
         addNotification({
